pretty InmemAppProxy
